requirements.txt
marvin
python-dotenv
TTS
ipykernel
pydub